requirements.txt
Django==5.1.4
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
psycopg2-binary==2.9.10
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.2
sqlparse==0.5.3
tzdata==2024.2
